blkiep

+++ canbushandling.go
@@ -115,7 +115,7 @@ func canPublish(cf CAN.CANFrame) {
 		log.Fatal(err)
 	}
 	if cf.ID > 2047 {
-		cf.ID &= ^(1 << 31)
+		cf.ID &= ^(uint32(1 << 31))
 	}
 
 	canSubscribe(cf.ID)
